Fixed the removing nonexistent port problem.

+++ client/js/Widgets/DiagramDesigner/ConnectionRouteManager3.js
@@ -109,7 +109,8 @@ define(['logManager', './AutoRouter', './Profiler'], function (logManager, AutoR
         this._onUnregisterSubcomponent = function(sender, ids){
             var longid = ids.objectID + DESIGNERITEM_SUBCOMPONENT_SEPARATOR + ids.subComponentID;
             try{
-                self.deleteItem(longid);
+                if(this._autorouterBoxes[longid])
+                    self.deleteItem(longid);
             }catch(e){
                 self.logger.error('ConnectionRouteManager3.deleteItem failed with error: ' + e);
             }
